% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anipose_2d.R
\docType{data}
\name{anipose_raw}
\alias{anipose_raw}
\title{Prices of 50,000 round cut diamonds}
\format{
A data frame with 53940 rows and 10 variables:
}
\usage{
anipose_raw
}
\description{
raw data
}
\details{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:

some details
}
\keyword{datasets}
