% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_poses.R
\name{augment_poses}
\alias{augment_poses}
\title{Augment poses}
\usage{
augment_poses(data, framerate, rollmean_frames)
}
\arguments{
\item{data}{Data frame}

\item{framerate}{Frame rate}

\item{rollmean_frames}{Number of frames for rolling mean}
}
\value{

}
\description{
Augment poses
}
