% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_likelihood.R
\name{filter_likelihood}
\alias{filter_likelihood}
\title{Filter likelihood}
\usage{
filter_likelihood(data, likelihood)
}
\arguments{
\item{data}{Data frame}

\item{likelihood}{Minimum likelihood}
}
\value{

}
\description{
Filter likelihood
}
