% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anipose.R
\name{tidy_anipose}
\alias{tidy_anipose}
\title{Pipe operator}
\usage{
tidy_anipose(data, likelihood = FALSE, interpolate = FALSE)
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
