% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anipose.R
\name{tidy_anipose}
\alias{tidy_anipose}
\title{Tidy Anipose}
\usage{
tidy_anipose(data, likelihood = FALSE, interpolate = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{likelihood}{Likelihood cut-off}

\item{interpolate}{Whether to interpolate or not (default = FALSE)}
}
\value{

}
\description{
Tidy Anipose
}
